Navigation to beerDetail fragment

+++ app/src/main/java/co/develhope/chooseyourownbeer/model/Beer.kt
@@ -1,6 +1,6 @@
 package co.develhope.chooseyourownbeer.model
 
-data class Beer(val id: Long,
+data class Beer(val id: Int,
                 var imagePath: Int,
                 var title: String,
                 var size: Double,
